Refactor Ratings component and add form submission functionality
